+++ Test/caseLED.h
@@ -7,15 +7,15 @@ class CaseLED{
     protected:
     //état de la case soit allume = true si LED allumée allume = false si éteinte
     bool allume;
-    uint8_t LedMat[2];
+    uint64_t LedMat[2];
 
     public:
     CaseLED(); //constructeur
     void allumer(); //fonction pour allumer la case
     void eteindre(); //fonction pour éteindre la case
     bool EstAllume(); //méthode pour connaître l'état de la LED
-    uint8_t getMat1();
-    uint8_t getMat2();
+    uint64_t getMat1();
+    uint64_t getMat2();
 };
